Combined selfprediction_model.py and downstream_model.py into model.py to remove redundancy.
